refactor(hook): modularize hook handling into dedicated module
Created src/hook module to encapsulate MLX hook logic (key events and main loop).

Moved draw.c, hook.c, and movement.c into src/hook.

Added hook.h header file to expose module interface.

Enhanced robustness and maintainability of hook.c and movement.c.

Introduced sort.c in src/hook for dedicated sprite sorting logic.

Updated include/cub3D.h and meson.build to reflect new structure.

+++ src/hook/hook.c
@@ -1,3 +1,4 @@
+#include "hook.h"
 #include "cub3D.h"
 
 void hook_key(mlx_key_data_t keydata, void *param) {
@@ -6,25 +7,21 @@ void hook_key(mlx_key_data_t keydata, void *param) {
 }
 
 void hook_generic(void *param) {
-    unsigned short fps = 1 / data.mlx->delta_time;
+    unsigned short fps = (unsigned short) (1 / data.mlx->delta_time);
     float movement = MOVEMENT / fps;
     float rotation = MOVEMENT / fps / 2;
 
     if (mlx_is_key_down(data.mlx, MLX_KEY_W))
         move((float[2]) {movement, 0});
-
     if (mlx_is_key_down(data.mlx, MLX_KEY_S))
         move((float[2]) {-movement, 0});
-
     if (mlx_is_key_down(data.mlx, MLX_KEY_A))
         move((float[2]) {0, movement});
-
     if (mlx_is_key_down(data.mlx, MLX_KEY_D))
         move((float[2]) {0, -movement});
 
     if (mlx_is_key_down(data.mlx, MLX_KEY_LEFT))
         rotate(-rotation);
-
     if (mlx_is_key_down(data.mlx, MLX_KEY_RIGHT))
         rotate(rotation);
 }
@@ -36,7 +33,7 @@ void hook_loop(void *param) {
     mlx_image_to_window(data.mlx, image, 0, 0);
     mlx_set_instance_depth(image->instances, 0);
 
-    float fov = (float) data.mlx->width / data.mlx->height / 2;
+    float fov = (float) data.mlx->width / (float) data.mlx->height / 2.0f;
     float plane[2] = {-fov * data.player.direction[1], fov * data.player.direction[0]};
     double distance_wall[image->width];
